pyproject.toml
[project]
name = "streamlit-bioroad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "googletrans==4.0.0rc1",
    "gtts>=2.5.4",
    "langchain>=0.2.16",
    "langchain-community>=0.2.16",
    "langchain-google-genai>=1.0.10",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.0",
    "ultralytics>=8.3.169",
]
